requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
googlemaps==4.10.0
grpcio==1.64.1
grpcio-tools==1.64.1
idna==3.7
protobuf==5.27.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.3.0
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
